edit を setting に統合 (#585)
# PRの概要

#572 の通り。

<img width="309" alt="image"
src="https://github.com/user-attachments/assets/7e85858c-118b-4b8a-aebe-c0bd0ee75f85"
/>


## 具体的な変更内容

edit/ を settings/ に統合

## 補足

関係ないエラー修正あり
https://github.com/ut-code/CourseMate/pull/585/commits/496d0170a2b693b1e4b13c10a91c801dcb18ad84


## レビューリクエストを出す前にチェック！

- [x] 改めてセルフレビューしたか
- [x] 手動での動作検証を行ったか
- [x] server の機能追加ならば、テストを書いたか
  - 理由: 書いた | server の機能追加ではない
- [x] 間違った使い方が存在するならば、それのドキュメントをコメントで書いたか
  - 理由: 書いた | 間違った使い方は存在しない
- [x] わかりやすいPRになっているか

<!-- レビューリクエスト後は、Slackでもメンションしてお願いすることを推奨します。 -->
